Created Rock_Paper_Scissors.py
Rock paper scissors - Created and edited to include more in-depth comments for a more tutorial look.
Hopefully, members of coding club will make their own version of rock paper scissors, either using this as a guide or by themselves. 👍
This may be the final product of many coding club members, but it is a good example of how python can be used to make a game - and how complicated you can make it if you want!